# openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getData",
        "parameters": [

        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/unsplash": {
      "get": {
        "operationId": "UnsplashController_getRandom",
        "summary": "get random unsplash picture",
        "parameters": [

        ],
        "responses": {
          "200": {
            "description": "The found record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unsplash"
                }
              }
            }
          }
        }
      }
    },
    "/unsplash/{id}": {
      "get": {
        "operationId": "UnsplashController_findOne",
        "summary": "find photo by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unsplash"
                }
              }
            }
          }
        }
      }
    },
    "/weather": {
      "get": {
        "operationId": "WeatherController_getCurrent",
        "summary": "get current weather by coords",
        "parameters": [
          {
            "name": "lon",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lang",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                }
              }
            }
          }
        }
      }
    },
    "/weather/{city}": {
      "get": {
        "operationId": "WeatherController_getCurrentByCity",
        "summary": "get current weather by city",
        "parameters": [
          {
            "name": "city",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "Percesso",
    "description": "The percesso API",
    "version": "1.0",
    "contact": {

    }
  },
  "tags": [
    {
      "name": "common",
      "description": ""
    }
  ],
  "servers": [

  ],
  "components": {
    "schemas": {
      "Urls": {
        "type": "object",
        "properties": {
          "full": {
            "type": "string"
          },
          "regular": {
            "type": "string"
          },
          "small": {
            "type": "string"
          },
          "thumb": {
            "type": "string"
          },
          "raw": {
            "type": "string"
          }
        },
        "required": [
          "full",
          "regular",
          "small",
          "thumb",
          "raw"
        ]
      },
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "city",
          "country"
        ]
      },
      "Links": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string"
          },
          "download": {
            "type": "string"
          }
        },
        "required": [
          "self",
          "download"
        ]
      },
      "Unsplash": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "example": {
              "created_at": "2016-05-03T11:00:28-04:00"
            }
          },
          "updated_at": {
            "type": "string",
            "example": {
              "updated_at": "2016-07-10T11:00:01-05:00"
            }
          },
          "urls": {
            "description": "contains all dimensions",
            "example": {
              "raw": "https://images.unsplash.com/photo-1417325384643-aac51acc9e5d",
              "full": "https://images.unsplash.com/photo-1417325384643-aac51acc9e5d?q=75&fm=jpg",
              "regular": "https://images.unsplash.com/photo-1417325384643-aac51acc9e5d?q=75&fm=jpg&w=1080&fit=max",
              "small": "https://images.unsplash.com/photo-1417325384643-aac51acc9e5d?q=75&fm=jpg&w=400&fit=max",
              "thumb": "https://images.unsplash.com/photo-1417325384643-aac51acc9e5d?q=75&fm=jpg&w=200&fit=max"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Urls"
              }
            ]
          },
          "author": {
            "description": "author picture",
            "example": {
              "id": "QPxL2MGqfrw",
              "username": "exampleuser"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ]
          },
          "location": {
            "description": "the location of the created photo",
            "example": {
              "name": "Montreal, Canada",
              "city": "Montreal",
              "country": "Canada"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Location"
              }
            ]
          },
          "links": {
            "description": "meta information about photo",
            "example": {
              "self": "https://api.unsplash.com/photos/Dwu85P9SOIk",
              "download": "https://unsplash.com/photos/Dwu85P9SOIk/download"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Links"
              }
            ]
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "urls",
          "author",
          "location",
          "links"
        ]
      },
      "Coord": {
        "type": "object",
        "properties": {
          "lon": {
            "type": "number"
          },
          "lat": {
            "type": "number"
          }
        },
        "required": [
          "lon",
          "lat"
        ]
      },
      "Wind": {
        "type": "object",
        "properties": {
          "speed": {
            "type": "number"
          },
          "deg": {
            "type": "number"
          },
          "gust": {
            "type": "number"
          }
        },
        "required": [
          "speed",
          "deg",
          "gust"
        ]
      },
      "Main": {
        "type": "object",
        "properties": {
          "temp": {
            "type": "number"
          },
          "feels_like": {
            "type": "number"
          },
          "temp_min": {
            "type": "number"
          },
          "temp_max": {
            "type": "number"
          },
          "pressure": {
            "type": "number"
          },
          "humidity": {
            "type": "number"
          },
          "sea_level": {
            "type": "number"
          },
          "grnd_level": {
            "type": "number"
          }
        },
        "required": [
          "temp",
          "feels_like",
          "temp_min",
          "temp_max",
          "pressure",
          "humidity",
          "sea_level",
          "grnd_level"
        ]
      },
      "Sys": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "type": "number"
          },
          "country": {
            "type": "string"
          },
          "sunrise": {
            "type": "number"
          },
          "sunset": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "country",
          "sunrise",
          "sunset",
          "name"
        ]
      },
      "Weather": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "coord": {
            "example": {
              "lon": 10.99,
              "lat": 44.34
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Coord"
              }
            ]
          },
          "base": {
            "type": "string",
            "example": "stations"
          },
          "visibility": {
            "type": "string",
            "example": 10000
          },
          "wind": {
            "example": {
              "speed": 0.62,
              "deg": 349,
              "gust": 1.18
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Wind"
              }
            ]
          },
          "rain": {
            "type": "object"
          },
          "clouds": {
            "type": "object"
          },
          "dt": {
            "type": "number"
          },
          "timezone": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "main": {
            "$ref": "#/components/schemas/Main"
          },
          "sys": {
            "$ref": "#/components/schemas/Sys"
          }
        },
        "required": [
          "id",
          "coord",
          "base",
          "visibility",
          "wind",
          "rain",
          "clouds",
          "dt",
          "timezone",
          "name",
          "main",
          "sys"
        ]
      }
    }
  }
}